login and auth changed
