+++ closet-tracker/app/(app)/(screens)/canvas.tsx
