+++ closet-tracker/app/(app)/(screens)/editItem.tsx
