+++ closet-tracker/app/(app)/(screens)/singleItem.tsx
