+++ closet-tracker/app/(app)/(tabs)/_layout.tsx
@@ -27,7 +27,7 @@ export default function TabLayout() {
       }}>
       
       <Tabs.Screen
-        name="home"
+        name="index"
         options={{
           title: 'Home',
           tabBarIcon: ({ color }) => <IconSymbol size={28} name="house.fill" color={color} />,
