+++ closet-tracker/app/(app)/(tabs)/index.tsx
