+++ closet-tracker/app/(app)/(tabs)/outfit.tsx
