+++ closet-tracker/app/(app)/(tabs)/upload.tsx
