+++ closet-tracker/app/(app)/(tabs)/user.tsx
@@ -79,7 +79,6 @@ export default function UserProfile() {
     try {
       await auth.signOut();
       console.log('User signed out successfully');
-      router.replace('/');
     } catch (error) {
       console.log('Error signing out: ', error);
     }
