+++ closet-tracker/app/(app)/(tabs)/wardrobe.tsx
