+++ closet-tracker/app/login.tsx
@@ -1,48 +1,27 @@
-import { Text, TextInput, ActivityIndicator, TouchableHighlight, StyleSheet } from 'react-native';
+import { View, Text, TextInput, ActivityIndicator, TouchableHighlight, StyleSheet, TouchableOpacity } from 'react-native';
 import React, { useState, useEffect } from 'react';
-import { createUserWithEmailAndPassword } from 'firebase/auth';
+import { signInWithEmailAndPassword } from 'firebase/auth';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { auth, db } from '@/FirebaseConfig';
+import { auth } from '@/FirebaseConfig';
 import { ScrollView } from 'react-native-gesture-handler';
-import { useRouter } from 'expo-router';
-import { doc, setDoc } from 'firebase/firestore';
+import { useRouter, Link } from 'expo-router';
 import beigeColors from '@/aesthetic/beigeColors';
+import { IconSymbol } from '@/components/ui/IconSymbol';
 
-export default function Signup() {
+export default function Login() {
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
-    const [confirmPassword, setConfirmPassword] = useState('');
-    const [name, setName] = useState('');
     const [loading, setLoading] = useState(false);
-
-    useEffect(() => {
-        const unsubscribe = auth.onAuthStateChanged((user) => {
-            if (user) {
-                router.replace('../(tabs)/wardrobe');
-            }
-        });
-        return unsubscribe;
-    }, []);
 
     const router = useRouter();
 
-    const signUp = async () => {
+    const signIn = async () => {
         setLoading(true);
         try {
-            if (password !== confirmPassword) {
-                alert("Passwords do not match.");
-                return;
-            }
-            const userCredential = await createUserWithEmailAndPassword(auth, email, password);
-            await setDoc(doc(db, 'users', userCredential.user.uid), {
-                name: name,
-                uid: userCredential.user.uid,
-                email: userCredential.user.email,
-            });
-
-            router.replace('../(tabs)/wardrobe');
+            await signInWithEmailAndPassword(auth, email, password);
+            router.replace('/');
         } catch (error: any) {
-            alert('Sign up failed: ' + error.message);
+            alert('Sign in failed: ' + error.message);
         } finally {
             setLoading(false);
         }
@@ -50,15 +29,16 @@ export default function Signup() {
 
     return (
         <SafeAreaView style={styles.container}>
-            <ScrollView contentContainerStyle={styles.scrollContainer}>
-                <Text style={styles.title}>Sign Up</Text>
-                <TextInput
-                    style={styles.input}
-                    placeholder="Name"
-                    placeholderTextColor="#FFFFFF"
-                    value={name}
-                    onChangeText={setName}
+            <Link href="/auth">
+                <IconSymbol
+                    name="chevron.left"
+                    size={30}
+                    color="#FFFFFF"
                 />
+            </Link>
+            <ScrollView contentContainerStyle={styles.scrollContainer}>
+                <Text style={styles.title}>Sign In</Text>
+
                 <TextInput
                     style={styles.input}
                     placeholder="Email"
@@ -68,6 +48,7 @@ export default function Signup() {
                     autoCapitalize="none"
                     keyboardType="email-address"
                 />
+
                 <TextInput
                     style={styles.input}
                     placeholder="Password"
@@ -77,19 +58,11 @@ export default function Signup() {
                     secureTextEntry
                 />
 
-                <TextInput
-                    style={styles.input}
-                    placeholder="Confirm Password"
-                    value={confirmPassword}
-                    onChangeText={setConfirmPassword}
-                    secureTextEntry
-                />
-
                 {loading ? (
                     <ActivityIndicator size="large" color={beigeColors.mutedGold} />
                 ) : (
-                    <TouchableHighlight style={styles.button} onPress={signUp}>
-                        <Text style={styles.buttonText}>Create Account</Text>
+                    <TouchableHighlight style={styles.button} onPress={signIn}>
+                        <Text style={styles.buttonText}>Login</Text>
                     </TouchableHighlight>
                 )}
             </ScrollView>
