+++ closet-tracker/app/signup.tsx
@@ -1,35 +1,40 @@
-import { View, Text, TextInput, ActivityIndicator, TouchableHighlight, StyleSheet } from 'react-native';
+import { Text, TextInput, ActivityIndicator, TouchableHighlight, StyleSheet, TouchableOpacity } from 'react-native';
 import React, { useState, useEffect } from 'react';
-import { signInWithEmailAndPassword } from 'firebase/auth';
+import { createUserWithEmailAndPassword } from 'firebase/auth';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { auth } from '@/FirebaseConfig';
+import { auth, db } from '@/FirebaseConfig';
 import { ScrollView } from 'react-native-gesture-handler';
-import { useRouter } from 'expo-router';
+import { useRouter, Link } from 'expo-router';
+import { doc, setDoc } from 'firebase/firestore';
 import beigeColors from '@/aesthetic/beigeColors';
+import { IconSymbol } from '@/components/ui/IconSymbol';
 
-export default function Login() {
+export default function Signup() {
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
+    const [confirmPassword, setConfirmPassword] = useState('');
+    const [name, setName] = useState('');
     const [loading, setLoading] = useState(false);
-
-    useEffect(() => {
-        const unsubscribe = auth.onAuthStateChanged((user) => {
-            if (user) {
-                router.replace('../(tabs)/wardrobe');
-            }
-        });
-        return unsubscribe;
-    }, []);
 
     const router = useRouter();
 
-    const signIn = async () => {
+    const signUp = async () => {
         setLoading(true);
         try {
-            const userCredential = await signInWithEmailAndPassword(auth, email, password);
-            router.replace('../(tabs)/wardrobe');
+            if (password !== confirmPassword) {
+                alert("Passwords do not match.");
+                return;
+            }
+            const userCredential = await createUserWithEmailAndPassword(auth, email, password);
+            await setDoc(doc(db, 'users', userCredential.user.uid), {
+                name: name,
+                uid: userCredential.user.uid,
+                email: userCredential.user.email,
+            });
+
+            router.replace('/');
         } catch (error: any) {
-            alert('Sign in failed: ' + error.message);
+            alert('Sign up failed: ' + error.message);
         } finally {
             setLoading(false);
         }
@@ -37,9 +42,22 @@ export default function Login() {
 
     return (
         <SafeAreaView style={styles.container}>
+            <Link href="/auth">
+                <IconSymbol
+                    name="chevron.left"
+                    size={30}
+                    color="#FFFFFF"
+                />
+            </Link>
             <ScrollView contentContainerStyle={styles.scrollContainer}>
-                <Text style={styles.title}>Sign In</Text>
-
+                <Text style={styles.title}>Sign Up</Text>
+                <TextInput
+                    style={styles.input}
+                    placeholder="Name"
+                    placeholderTextColor="#FFFFFF"
+                    value={name}
+                    onChangeText={setName}
+                />
                 <TextInput
                     style={styles.input}
                     placeholder="Email"
@@ -49,7 +67,6 @@ export default function Login() {
                     autoCapitalize="none"
                     keyboardType="email-address"
                 />
-
                 <TextInput
                     style={styles.input}
                     placeholder="Password"
@@ -59,11 +76,19 @@ export default function Login() {
                     secureTextEntry
                 />
 
+                <TextInput
+                    style={styles.input}
+                    placeholder="Confirm Password"
+                    value={confirmPassword}
+                    onChangeText={setConfirmPassword}
+                    secureTextEntry
+                />
+
                 {loading ? (
                     <ActivityIndicator size="large" color={beigeColors.mutedGold} />
                 ) : (
-                    <TouchableHighlight style={styles.button} onPress={signIn}>
-                        <Text style={styles.buttonText}>Login</Text>
+                    <TouchableHighlight style={styles.button} onPress={signUp}>
+                        <Text style={styles.buttonText}>Create Account</Text>
                     </TouchableHighlight>
                 )}
             </ScrollView>
