+++ closet-tracker/context/UserContext.tsx
@@ -1,0 +1,32 @@
+import React, { createContext, useState, useEffect, useContext } from 'react';
+import { getAuth, onAuthStateChanged, User } from 'firebase/auth';
+
+interface UserContextProps {
+  currentUser: User | null;
+}
+
+const UserContext = createContext<UserContextProps>({ currentUser: null });
+
+interface UserProviderProps {
+  children: React.ReactNode;
+}
+
+export const UserProvider: React.FC<UserProviderProps> = ({ children }) => {
+  const [currentUser, setCurrentUser] = useState<User | null>(null);
+  const auth = getAuth();
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, (user) => {
+      setCurrentUser(user);
+    });
+    return unsubscribe;
+  }, [auth]);
+
+  return (
+    <UserContext.Provider value={{ currentUser }}>
+      {children}
+    </UserContext.Provider>
+  );
+};
+
+export const useUser = () => useContext(UserContext);
